update task controller

+++ app/Http/Controllers/Api/TaskController.php
@@ -55,37 +55,59 @@ class TaskController extends Controller
     public function updateTaskToComplete($id)
     {
         $task = Task::find($id);
-        $task->task_desc = $task->task_desc;
-        $task->task_is_active = false;
-        $task->task_is_completed = true;
-        $task->updated_at = date('Y-m-d H:i:s');
-        $task->save();
 
-        $response = [
-            'status'    => Response::HTTP_OK,
-            'message'   => 'Task marked in complete list',
-            'data'      => array(),
-        ];
+        if (!empty($task)) {
+            $task->task_desc = $task->task_desc;
+            $task->task_is_active = false;
+            $task->task_is_completed = true;
+            $task->updated_at = date('Y-m-d H:i:s');
+            $task->save();
 
-        return response()->json($response, Response::HTTP_OK);
+            $response = [
+                'status'    => Response::HTTP_OK,
+                'message'   => 'Task marked in complete list',
+                'data'      => array(),
+            ];
+
+            return response()->json($response, Response::HTTP_OK);
+        } else {
+            $response = [
+                'status'    => Response::HTTP_BAD_REQUEST,
+                'message'   => 'Task List is empty',
+                'data'  => array(),
+            ];
+
+            return response()->json($response, Response::HTTP_BAD_REQUEST);
+        }
     }
 
     public function updateTaskToActive($id)
     {
         $task = Task::find($id);
-        $task->task_desc = $task->task_desc;
-        $task->task_is_active = true;
-        $task->task_is_completed = false;
-        $task->updated_at = date('Y-m-d H:i:s');
-        $task->save();
 
-        $response = [
-            'status'    => Response::HTTP_OK,
-            'message'   => 'Task marked in active list',
-            'data'      => array(),
-        ];
+        if (!empty($task)) {
+            $task->task_desc = $task->task_desc;
+            $task->task_is_active = true;
+            $task->task_is_completed = false;
+            $task->updated_at = date('Y-m-d H:i:s');
+            $task->save();
 
-        return response()->json($response, Response::HTTP_OK);
+            $response = [
+                'status'    => Response::HTTP_OK,
+                'message'   => 'Task marked in active list',
+                'data'      => array(),
+            ];
+
+            return response()->json($response, Response::HTTP_OK);
+        } else {
+            $response = [
+                'status'    => Response::HTTP_BAD_REQUEST,
+                'message'   => 'Task List is empty',
+                'data'  => array(),
+            ];
+
+            return response()->json($response, Response::HTTP_BAD_REQUEST);
+        }
     }
 
     public function updateAllTaskToComplete()
@@ -189,15 +211,26 @@ class TaskController extends Controller
     public function deleteTask($id)
     {
         $task = Task::find($id);
-        $task->delete();
 
-        $response = [
-            'status'    => Response::HTTP_OK,
-            'message'   => 'Task deleted',
-            'data'  => array(),
-        ];
+        if (!empty($task)) {
+            $task->delete();
 
-        return response()->json($response, Response::HTTP_OK);
+            $response = [
+                'status'    => Response::HTTP_OK,
+                'message'   => 'Task deleted',
+                'data'  => array(),
+            ];
+
+            return response()->json($response, Response::HTTP_OK);
+        } else {
+            $response = [
+                'status'    => Response::HTTP_BAD_REQUEST,
+                'message'   => 'Task List is empty',
+                'data'  => array(),
+            ];
+
+            return response()->json($response, Response::HTTP_BAD_REQUEST);
+        }
     }
 
     public function deleteAllCompleteTask()
